Crate tidb_query_codegen

Procedural macros used in the tidb_query component; part of the coprocessor subsystem.

For an overview of the coprocessor architecture, see the documentation on tikv/src/coprocessor.

This crate exports a custom derive for AggrFunction and an attribute macro called rpn_fn for use on functions which provide coprocessor functionality. rpn_fn is documented in the rpn_function module.

Modules

aggr_function

rpn_function
Implementation of the rpn_fn attribute macro.

Attribute Macros

rpn_fn
The rpn_fn attribute.

Derive Macros

AggrFunction
The AggrFunction custom derive.